backend/requirements.txt
attr==0.3.2
brotli==1.2.0
brotlicffi==1.2.0.1
chardet==7.4.3
click==8.3.3
cloudinary==1.44.2
ConfigParser==7.2.0
cryptography==47.0.0
cython==3.2.4
defusedxml==0.7.1
docutils==0.22.4
email_validator==2.3.0
eval_type_backport==0.3.1
exceptiongroup==1.3.1
fastapi==0.136.1
filelock==3.29.0
filetype==1.2.0
grandalf==0.8
h2==4.3.0
HTMLParser==0.0.2
hypothesis==6.152.2
ipython==9.13.0
ipywidgets==8.1.8
Jinja2==3.1.6
jnius==1.1.0
js==1.0
keyring==25.7.0
langchain_anthropic==1.4.1
langchain_google_genai==4.2.2
langchain_openai==1.2.1
langchain_text_splitters==1.1.2
langgraph_api==0.8.1
langsmith_pyo3==0.1.0rc5
numpy==2.4.4
outcome==1.3.0.post0
pandas==3.0.2
protobuf==7.34.1
pycryptodome==3.23.0
pygraphviz==1.14
pyodide==0.0.2
pyOpenSSL==26.1.0
pyppeteer==2.0.0
pytest==9.0.3
python-dotenv==1.2.2
rapidfuzz==3.14.5
simplejson==4.1.1
simsimd==6.5.16
sniffio==1.3.1
socksio==1.0.0
Sphinx==9.1.0
starlette==1.0.0
streamlit==1.56.0
toml==0.10.2
tornado==6.5.5
tqdm==4.67.3
urllib3_secure_extra==0.1.0
uvloop==0.22.1
websockets==16.0
winloop==0.6.2
wrapt==2.1.2
xmlrpclib==1.0.1
xxhash==3.7.0
